w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/text-level-semantics/the-a-element/a.text-getter-01-expected.txt

a b c a b c a b c a c a b c

PASS HTMLAnchorElement.text getting 
PASS Test for anchor 0 
PASS Test for anchor 1 
PASS Test for anchor 2 
FAIL Test for anchor 3 assert_equals: expected "a b c " but got "a c "(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:29:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:28:9
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:24:5)
FAIL Test for anchor 4 assert_equals: expected "a b c " but got "a b c"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:29:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:28:9
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/text-level-semantics/the-a-element/a.text-getter-01.html:24:5)
